+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # develop
 
 ## New features
+- BDN-1070 Added Yandex Bidding support
 - BND-1046 Networks update
 - APDM-1631 Podspec generation - update
 - BDN-1083 CI S3 zip-archive checking